# src/main/resources/openapi/api.yaml
openapi: 3.0.0
info:
  title: TODO API
  description: シンプルなTODO管理API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 開発サーバー
tags:
  - name: todos
    description: TODO操作

paths:
  /todos:
    get:
      tags:
        - todos
      summary: TODO一覧を取得
      operationId: getTodos
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      tags:
        - todos
      summary: 新しいTODOを作成
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: 不正なリクエスト

  /todos/{todoId}:
    get:
      tags:
        - todos
      summary: 特定のTODOを取得
      operationId: getTodoById
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: TODOが見つからない
    put:
      tags:
        - todos
      summary: TODOを更新
      operationId: updateTodo
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: TODOが見つからない
    delete:
      tags:
        - todos
      summary: TODOを削除
      operationId: deleteTodo
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 削除成功
        '404':
          description: TODOが見つからない

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: TODO ID
        title:
          type: string
          description: TODOのタイトル
        description:
          type: string
          description: TODOの詳細説明
        completed:
          type: boolean
          description: 完了フラグ
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時

    TodoCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: TODOのタイトル
        description:
          type: string
          maxLength: 500
          description: TODOの詳細説明

    TodoUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: TODOのタイトル
        description:
          type: string
          maxLength: 500
          description: TODOの詳細説明
        completed:
          type: boolean
          description: 完了フラグ 